Added missing test when not using get()
Test that all features are covered w/o going through get() functionality
and are thus covered by other tests (that should check correctness)

* fix test_feature_compat docstring
* move ichain_2 to tests as it wasn't being used elsewhere.  This may
  get moved back
